requirements.txt
flask
Flask-CORS
pandas
numpy
ipykernel
pinecone
python-dotenv
sentence-transformers
matplotlib
seaborn
pytesseract
beautifulsoup4
requests
opencv-python
opencv-python-headless 
pillow
tqdm
Google-Images-Search
windows-curses
ipywidgets
transformers
easyocr
google-api-core 
google-api-python-client
google-auth
google-auth-httplib2 
Google-Images-Search
selenium
torch
torchvision
scikit-image
aiohttp
aioredis
cachetools
tenacity
pytest
pytest-flask
requests
pytest-asyncio
mypy
pyspellchecker
google-generativeai
autoflake
ruff